+++ employ-easy/src/main/java/se/yrgo/employee/repositories/EmployeeRepository.java
@@ -1,6 +1,8 @@
-package se.yrgo.employee;
+package se.yrgo.employee.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import se.yrgo.employee.domain.Employee;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Long> {
