+++ employ-easy/src/main/java/se/yrgo/employee/services/EmployeeService.java
@@ -1,11 +1,14 @@
-package se.yrgo.employee;
+package se.yrgo.employee.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.GetMapping;
 
+import se.yrgo.employee.domain.Employee;
+import se.yrgo.employee.repositories.EmployeeRepository;
 
 @Service
 public class EmployeeService {
@@ -16,7 +19,7 @@ public class EmployeeService {
 	public EmployeeService(EmployeeRepository employeeRepository) {
 		this.employeeRepository = employeeRepository;
 	}
-	
+
 	@GetMapping
 	public List<Employee> findAll() {
 		return employeeRepository.findAll();
@@ -24,6 +27,10 @@ public class EmployeeService {
 
 	public void addEmployee(Employee employee) {
 		employeeRepository.save(employee);
-		
+	}
+
+	public Optional<Employee> findById(Long id) {
+		Optional<Employee> employee = employeeRepository.findById(id);
+		return employee;
 	}
 }
