+++ employ-easy/src/main/resources/application.properties
@@ -1,6 +1,7 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/employees
-spring.datasource.username=
-spring.datasource.password=
+spring.datasource.username=postgres
+spring.datasource.password=Borealis16
+
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
